requirements.txt
pandas>=1.3.0
tabulate>=0.8.9
requests>=2.26.0
beautifulsoup4>=4.9.3
python-dotenv>=0.19.0
SQLAlchemy>=1.4.0 